refactor: split inbox_handler.py into focused modules (#972)
Extract port-factory wiring and pre-bootstrap pending-queue management
into dedicated submodules:

- inbox_port_factories.py: _sync_port_factory, _trigger_activity_port_factory,
  _sync_and_trigger_port_factory, plus the three semantics-set constants
- inbox_pending_queue.py: _activity_context_id, _queue_pending_case_activity,
  _expire_pending_case_activities, _replay_pending_case_activities

inbox_handler.py is reduced from 666 to 413 lines; it re-exports all moved
names so existing callers and monkeypatching tests continue to work.

inbox_pipeline.py updated to import _activity_context_id and
_queue_pending_case_activity from inbox_pending_queue instead of defining
local duplicates, removing the only code duplication between the two modules.
